+++ RELEASE-NOTES.md
@@ -1,3 +1,16 @@
+# 2.39.0
+* Support for "strict" operation matching that strictly matches trailing slashes on a path
+  [[Details]](https://bitbucket.org/atlassian/swagger-request-validator/pull-requests/405)
+
+# 2.38.1
+* Support operation response code ranges
+  [[Details]](https://bitbucket.org/atlassian/swagger-request-validator/pull-requests/404)
+
+# 2.38.0
+* Support configuring the schema validation cache size. Allows consumers to limit the size to avoid OOME.
+  [[#774]](https://bitbucket.org/atlassian/swagger-request-validator/issues/774)
+  [[Details]](https://bitbucket.org/atlassian/swagger-request-validator/pull-requests/395)
+
 # 2.37.0
 * Bump version of `pact` to `4.6.0`
   [[#772]](https://bitbucket.org/atlassian/swagger-request-validator/issues/772)